refactor: replace getClient to getClientPublic
