+++ src/features/oauth/index.ts
@@ -2,6 +2,6 @@ export * from './components/authorize-form';
 export * from './frames/authorize-frame';
 export * from './hooks/use-authorize-form';
 export * from './hooks/use-client';
-export * from './services/get-client';
+export * from './services/get-client-public';
 export * from './services/get-clients';
 export * from './services/consent';
