+++ src/routes/_auth-required/authorize.tsx
@@ -5,7 +5,7 @@ import { AuthorizeFrame } from '@/features/oauth';
 
 const schema = z.object({
   clientId: z.string().min(1, 'Client ID is required'),
-  redirectUrl: z.string(),
+  redirectUrl: z.string().url('Invalid redirect URL'),
 });
 
 const AuthorizePage = () => {
